bink: add gas cost
## Introduces a new type, +bone, a bounded +tone

For +bone, there's a compromise between tightness of type and terseness
of translation from +mink to +bink.

The ideal-ish type for this seems like it would be:

    +$  bone
      $@  ~
      $:  tone
          rem=@
      ==

I.e., represent 'out-of-gas' as the atom `~`, and non-out-of-gas as a
pair of `+tone` and gas remaining. This, alas, breaks because `=^` needs
`r` to be a cell in all cases and has no fall-through for an atomic
result.

Another possibility, that would cut down the number of lines in the
translation from +mink to +bink, would be:

    +$  bone
      $:  $%  [%0 ...]
              ...
              [%3 ~]
          $=  rem  @
      ==

I.e., write out the cases of +tone by hand, and add another arm to
the $% for 'out-of-gas'. With this, we could leave alone the various
instances of e.g. `?.  ?=(%0 -.head)  head^bud`. But this is kind of
gross; it's using a lot of information density to convey 'nothing'.

We settled upon `$:($@(~ tone) rem=@)`. There were two choices for the
various checks against the head of the +tone: either change from e.g.
`?=(%0 -.head)` to `?=([%0 *] head)` in all cases, or add an extra
`?~  head  ~` prior to the check. Opted for the latter one.

## Adds naive gas cost computation

The whole point of +bink is that it's a +mink with a gas function. There
are two places where gas costs may make sense: noun construction and
computation steps. We supply extremely naive functions for both. For
noun construction, we say that each 256 bits of atom size is of cost 1,
and each cell is cost 1 plus the cost of its head plus the cost of its
tail. For computation, we assume every computational step has cost 1.

This is just to get a starting point; we will later on refine this based
on measurements, handle edge cases like nock 4 (which may change its
input's size), and handle jets.