dao: use list of updates in proposals
Each proposal previously contained one update
to run if the proposal passed.
Now, each proposal contains a list of updates
to be run in order if the proposal passes.
This allows, e.g., a host change to be accomplished
in one, rather than two or more, proposals.

The advantage here is two-fold:
1. Reduce cognitive load on DAO participants by
   bundling connected actions together.
2. Reduce the cost to DAO participants by reducing
   the number of votes required by bundling actions.