multisig: flatten code structure a bit
This change was purely for understanding, in a way.
The old way makes sense only after I made this modification.
IMHO, less nesting/extraction makes the code dumber and slightly more followable,
at the cost of repetition (which very well may lead to mistakes)

Open to reverting this change at a later date.